requirements.txt
aerich==0.7.1
aiohttp==3.8.1
aiosignal==1.2.0
aiosqlite==0.17.0
appdirs==1.4.4
APScheduler==3.8.1
async-timeout==4.0.1
asyncpg==0.25.0
attrs==21.2.0
base58==2.1.1
bitarray==1.2.2
black==22.6.0
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.7
click==8.0.3
click-aliases==1.0.1
cytoolz==0.11.2
dictdiffer==0.9.0
distlib==0.3.5
ecdsa==0.17.0
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
filelock==3.8.0
frozenlist==1.2.0
hdwallet==2.0.1
hexbytes==0.2.2
identify==2.5.3
idna==3.3
iso8601==0.1.16
jedi==0.18.1
jsonschema==3.2.0
lru-dict==1.1.7
marshmallow==3.14.1
marshmallow-dataclass==8.5.3
mnemonic==0.20
multiaddr==0.0.9
multidict==5.2.0
mypy-extensions==0.4.3
netaddr==0.8.0
nodeenv==1.7.0
parsimonious==0.8.1
parso==0.8.2
pathspec==0.9.0
platformdirs==2.5.2
pre-commit==2.20.0
protobuf==3.19.1
psutil==5.9.1
ptpython==3.0.23
pycryptodome==3.11.0
pydantic==1.9.2
Pygments==2.10.0
pypika-tortoise==0.1.1
pyrsistent==0.18.0
pysha3==1.0.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.26.0
rlp==2.0.1
six==1.16.0
tabulate==0.8.9
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
toolz==0.11.2
tortoise-cli==0.1.1
tortoise-orm==0.17.8
typing-inspect==0.7.1
typing_extensions==4.0.0
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
varint==1.0.2
virtualenv==20.16.3
wcwidth==0.2.5
web3 @ git+https://github.com/Rock-n-Block/web3.py.git@80dd4f00cd7274b576ee0e09364ed47a4de09ea9
websockets==10.1
yarl==1.7.2
fastapi==0.97.0
uvicorn==0.22.0
